requirements.txt
playwright~=1.31.1
pytest~=7.2.2
Pillow~=10.0.1
pytest-html==4.1.1